services/voice/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.2.0
grpcio>=1.80.0
grpcio-tools>=1.80.0
python-jose[cryptography]>=3.3.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
pytest>=8.0.0
websockets>=12.0.0
